Make new hardware contexts compile
