--- lib/controls/hardware/context.h
@@ -1,13 +1,0 @@
-#include <memory>  // For shared_ptr
-
-#include "controls/button.h"
-
-namespace controls::hardware {
-
-class Context {
-  public:
-    virtual bool PollAll() = 0;
-    virtual std::shared_ptr<Button> CreateButon(uint8_t pin1, uint8_t pin2 = 0) = 0;
-};
-
-}  // namespace controls::hardware
