+++ src/GliderWars.cpp
@@ -5,7 +5,7 @@
 // #include <H2HDisplay.h>
 // #include <Game.h>
 // #include <GliderWars.h>
-// #include <simple.h>
+// #include <hardware_simple.h>
 // #include <H2HControl.h>
