+++ src/TwentyTest.cpp
@@ -1,26 +1,26 @@
 
-#include <Arduino.h>
-#include <FastLED.h>
-#include <Display.h>
-#include <TwentyDisplay.h>
-#include <Game.h>
-#include <NoiseGame.h>
+// #include <Arduino.h>
+// #include <FastLED.h>
+// #include <Display.h>
+// #include <TwentyDisplay.h>
+// #include <Game.h>
+// #include <NoiseGame.h>
 
 
-Game* game;
-Display* gameDisplay;
+// Game* game;
+// Display* gameDisplay;
 
-void setup()
-{
-    gameDisplay = (Display*)new TwentyDisplay();
+// void setup()
+// {
+//     gameDisplay = (Display*)new TwentyDisplay();
 
-    game = (Game*)new NoiseGame(gameDisplay);
-    game->setup();
-    FastLED.show();
-}
+//     game = (Game*)new NoiseGame(gameDisplay);
+//     game->setup();
+//     FastLED.show();
+// }
 
-void loop()
-{
-    game->loop();
-    FastLED.show();
-}
+// void loop()
+// {
+//     game->loop();
+//     FastLED.show();
+// }
